Cleanup: rm nonparallel.jl, common.jl
rename parallel.jl to auxfunctions.jl
move the common code in there
