+++ src/auxfunctions.jl
@@ -4,8 +4,6 @@ using LinearAlgebra
 using Base.Threads
 # using Distributed
 
-include("common.jl")
-
 export compute_weights, compute_DI, compute_FN, remove_duplicate_seqs
 
 ## function borrowed from stdlib/Distributed test file
@@ -41,6 +39,176 @@ function disable_blas_threads(f)
         LinearAlgebra.BLAS.set_num_threads(old_num_threads)
     end
 end
+
+const PACKBITS = 64
+const _u = 0x0084210842108421
+const _alt = 0x007c1f07c1f07c1f
+
+const packfactor = PACKBITS ÷ 5
+const packrest = PACKBITS % 5
+
+const _msk = ~(UInt64(0));
+
+clength(l::Int) = (l-1) ÷ packfactor + 1
+crest(l::Int)   = (l-1) % packfactor + 1
+
+nnz_aux(x::UInt64) = ((x | (x >>> 1) | (x >>> 2) | (x >>> 3) | (x >>> 4)) & _u)
+
+nz_aux(nx::UInt64) = (nx & (nx >>> 1) & (nx >>> 2) & (nx >>> 3) & (nx >>> 4) & _u)
+
+nz_aux2(nx::UInt64, s) = nz_aux(nx) & (_msk >>> s)
+
+@inline function collapse(z)
+    t = (z & _alt) + ((z >> 5) & _alt)
+    t += t >> 10
+    t += t >> 20
+    t += t >> 40
+    t &= 0x3ff
+    return t
+end
+
+function compress_Z(Z::Matrix{Int8})
+    N, M = size(Z)
+    ZZ = Vector{Int8}[Z[:,i] for i = 1:M]
+
+    cl = clength(N)
+
+    cZ = [zeros(UInt64, cl) for i=1:M]
+
+    @inbounds for i = 1:M
+        cZi = cZ[i]
+        ZZi = ZZ[i]
+        for k = 1:N
+            k0 = (k-1) ÷ packfactor + 1
+            k1 = (k-1) % packfactor
+            cZi[k0] |= (UInt64(ZZi[k]) << (5 * k1))
+        end
+    end
+
+    return cZ
+end
+
+macro hash(x)
+    tst = get(ENV, "GDCA_TESTING", "false") == "true"
+    return Expr(:call, tst ? :sum : :hash, esc(x))
+end
+
+function remove_duplicate_seqs(Z::Matrix{Int8})
+    N, M = size(Z)
+    hZ = Array{UInt}(undef, M)
+    @inbounds for i = 1:M
+        hZ[i] = @hash(@view Z[:,i])
+    end
+    print("removing duplicate sequences... ")
+
+    ref_seq_ind = Array{Int}(undef, M)
+    ref_seq = Dict{UInt,Int}()
+    @inbounds for i = 1:M
+        ref_seq_ind[i] = get!(ref_seq, hZ[i], i)
+    end
+    uniqueseqs = collect(values(ref_seq))
+
+    # Check for collisions
+    old_collided = trues(M)
+    collided = falses(M)
+    while true
+        fill!(collided, false)
+        @inbounds for i = 1:M
+            k = ref_seq_ind[i]
+            (!old_collided[i] || k == i) && continue
+            collided[i] = @views Z[:,i] ≠ Z[:,k]
+        end
+        any(collided) || break
+
+        # Collect index of first row for each collided hash
+        empty!(ref_seq)
+        @inbounds for i = 1:M
+            collided[i] || continue
+            ref_seq_ind[i] = get!(ref_seq, hZ[i], i)
+        end
+        for v in values(ref_seq)
+            push!(uniqueseqs, v)
+        end
+        old_collided, collided = collided, old_collided
+    end
+    sort!(uniqueseqs)
+
+    newM = length(uniqueseqs)
+    newZ = Z[:,uniqueseqs]
+    println("done: $M -> $newM")
+    return newZ, uniqueseqs
+end
+
+compute_weights(Z::Matrix{Int8}, θ) = compute_weights(Z, maximum(Z), θ)
+
+compute_weights(Z::Matrix{Int8}, q, θ) = throw(ArgumentError("θ must be either :auto or a single real value"))
+
+function Z_to_cZ(Z, q)
+    M = size(Z, 2)
+    fast = q ≤ 32 && get(ENV, "GDCA_FORCE_FALLBACK", "false") ≠ "true"
+    fast || println("GaussDCA: using slower fallbacks")
+    return fast ? compress_Z(Z) : [Z[:,i]::Vector{Int8} for i = 1:M]
+end
+
+function compute_weights(Z::Matrix{Int8}, q, θ::Symbol)
+    θ ≠ :auto && throw(ArgumentError("θ must be either :auto or a single real value"))
+
+    N, M = size(Z)
+    cZ = Z_to_cZ(Z, q)
+    θ = compute_theta(cZ, N, M)
+    return compute_weights(cZ, θ, N, M)
+end
+
+function compute_weights(Z::Matrix{Int8}, q, θ::Real)
+    N, M = size(Z)
+    cZ = Z_to_cZ(Z, q)
+    return compute_weights(cZ, θ, N, M)
+end
+
+function compute_FN(mJ::Matrix{Float64}, N::Int, q::Integer)
+    q = Int(q)
+    s = q - 1
+
+    mJij = Array{Float64}(undef, s, s)
+    amJi = Array{Float64}(undef, s)
+    amJj = Array{Float64}(undef, s)
+    fs = Float64(s)
+    fs2 = Float64(s^2)
+
+    FN = zeros(N, N)
+
+    for i = 1:N-1
+        row0 = (i-1) * s
+        for j = i+1:N
+            col0 = (j-1) * s
+
+            ## devectorize for speed and memory efficicency
+            # mJij = mJ[row,col]
+            # mK = mJij .- mean(mJij, 1) .- mean(mJij, 2) .+ mean(mJij)
+
+            amJ = 0.0
+            fill!(amJi, 0.0)
+            fill!(amJj, 0.0)
+            for b = 1:s, a = 1:s
+                x = mJ[row0 + a, col0 + b]
+                mJij[a,b] = x
+                amJi[b] += x / fs
+                amJj[a] += x / fs
+                amJ += x / fs2
+            end
+            fn = 0.0
+            for b = 1:s, a = 1:s
+                fn += (mJij[a,b] - amJi[b] - amJj[a] + amJ)^2
+            end
+
+            FN[i,j] = √fn
+            FN[j,i] = FN[i,j]
+        end
+    end
+    return FN
+end
+
+## parallel code
 
 const TriuInd = Tuple{Tuple{Int,Int},Tuple{Int,Int},Int}
